requirements.txt
# YouTube Recipe Generator - Python Dependencies

# Google Gemini API
google-generativeai>=0.3.2

# JSON Schema validation
jsonschema>=4.20.0

# HTTP requests
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Note: tkinter usually comes with Python, but if not installed:
# On Ubuntu/Debian: sudo apt-get install python3-tk
# On macOS: usually included with Python
# On Windows: usually included with Python
